tests working after refactor
