+++ superset-frontend/package.json
@@ -148,7 +148,6 @@
     "html-webpack-plugin": "^5.3.2",
     "immer": "^9.0.6",
     "interweave": "^13.0.0",
-    "jest-canvas-mock": "^2.5.2",
     "jquery": "^3.5.1",
     "js-levenshtein": "^1.1.6",
     "js-yaml-loader": "^1.2.2",
@@ -316,6 +315,7 @@
     "ignore-styles": "^5.0.1",
     "imports-loader": "^3.1.1",
     "jest": "^26.6.3",
+    "jest-canvas-mock": "^2.5.2",
     "jest-environment-enzyme": "^7.1.2",
     "jest-enzyme": "^7.1.2",
     "jest-websocket-mock": "^2.2.0",
